etl/requirements.txt
numpy
seaborn
matplotlib
scipy
pandas
mysql-connector-python